% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize recording data based on dates/times}
\usage{
categorize(data, category.name, start.date, end.date, start.time, end.time)
}
\arguments{
\item{data}{Selection data.frame.}

\item{category.name}{String of what you want to call your category. Example: "Early Morning".}

\item{start.date}{Start date for given category. Defaults to min date. Accepts multiple inputs including column name of a calculated field. Has to be numeric yday, string in "YYYY-MM-DD" format, Date format, or POSIXct.}

\item{end.date}{End date for given category. Defaults to max date. Accepts multiple inputs including column name of a calculated field. Has to be numeric yday, string in "YYYY-MM-DD" format, Date format, or POSIXct.}

\item{start.time}{Start time for a given category. Defaults to min time. Accepts multiple inputs including column name of a calculated field. Has to be a string in "hh:mm:ss" format, hms format, or POSIXct.}

\item{end.time}{End time for a given category. Defaults to max time. Accepts multiple inputs including column name of a calculated field. Has to be a string in "hh:mm:ss" format, hms format, or POSIXct.}
}
\value{
A data.frame with a category column of defined categories.
}
\description{
Categorize recording data based on dates/times
}
\examples{
data \%>\%  dplyr::group_by(location) \%>\% dplyr::mutate(category=NA) \%>\% categorize("EN","2017-16-01","2017-16-15","22:00:00","23:59:00")
}
\keyword{categorize,}
\keyword{get,}
\keyword{recordings}
\keyword{wav,}
