% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwavs.R
\name{get.wavs}
\alias{get.wavs}
\title{Get .wav Recording Metadata}
\usage{
get.wavs(
  directory = getwd(),
  start.date = NULL,
  end.date = NULL,
  timezone = "America/Denver",
  getDuration = TRUE,
  minduration = 0
)
}
\arguments{
\item{directory}{String file path to folder containing location folders of recording data. Defualts to working directroy.}

\item{start.date}{String start date for when to gather recordings from in YYYY-MM-DD format. For example "2017-06-01"}

\item{end.date}{String end date for when to gather recordings till in YYYY-MM-DD format. For example ""2017-06-30"}

\item{timezone}{Timezone to have datetime column of output in. Defaults to "America/Denver" where this package was developed.}

\item{getDuration}{Binary true or false on whether to calculate file duration. This process can take time and is optional but is needed to exclude recordings that are under a certain length.}

\item{minduration}{Minimum duration in seconds of the recordings to read in. This will only apply when getDuration=T.}
}
\value{
A data.frame with recording file metadata: file.path, basename, location, datetime, JDay, and sometimes file.duration.
}
\description{
This function collects and compiles recording metadata into a usable data.frame for recording selection based of a file directory.
}
\examples{
data <- get.wavs(start.date = "2017-06-01", end.date = "2017-06-30", getDuration=T, minduration=180) #gets recordings of at least 3 min long for the month of June 2o17.
}
\keyword{get,}
\keyword{recordings}
\keyword{wav,}
