% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSunCalcs.R
\name{getSunCalcs}
\alias{getSunCalcs}
\title{Get sunlight times for recording selection}
\usage{
getSunCalcs(
  data,
  calc = c("solarNoon", "nadir", "sunrise", "sunset", "sunriseEnd", "sunsetStart",
    "dawn", "dusk", "nauticalDawn", "nauticalDusk", "nightEnd", "night", "goldenHourEnd",
    "goldenHour"),
  doParallel = FALSE
)
}
\arguments{
\item{data}{Selection data.frame as created by the getwavs() function. Must contain existing datetime, Latitude, and Longitude columns to compute.}

\item{calc}{The type of calculations to derive. Default is all: c("solarNoon", "nadir", "sunrise", "sunset", "sunriseEnd","sunsetStart", "dawn", "dusk", "nauticalDawn", "nauticalDusk", "nightEnd", "night", "goldenHourEnd", "goldenHour")}

\item{doParallel}{Binary TRUE or FALSE whether to run process in parallel to save time. Will work better on some machines vs others.}
}
\value{
A data.frame with appended columns of sun times as selected.
}
\description{
This function appends columns of sun times for recording selection such as sunrise/sunset.
}
\examples{
data <- getSunCalcs(data, calc = c("sunrise","sunset"))
}
\keyword{get,}
\keyword{recordings}
\keyword{suncalc,}
