% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergelocations.R
\name{mergelocations}
\alias{mergelocations}
\title{Merge location and recording data}
\usage{
mergelocations(
  data = NULL,
  locationdata = NULL,
  locationname = "location",
  Latitude = "Latitude",
  Longitude = "Longitude"
)
}
\arguments{
\item{data}{Selection data.frame as created by the getwavs() function. Must contain the location column for merging.}

\item{locationdata}{Location data.frame containing the locations that match with recording selection and their coordinates.}

\item{locationname}{String column name for locations in your locationdata file if differs from default: location.}

\item{Latitude}{String column name for Latitude in your locationdata file if differs from default: Latitude.}

\item{Longitude}{String column name for Longitude in your locationdata file if differs from default: Longitude.}
}
\description{
This function introduces a simplified merge for recording data and location data. Regular merge methods can be destructive and therefore impact recording selection. This method provides checks to prevent this from happening.
}
\examples{
data <- mergelocations(data,locationdata,locationname = "Location")
}
\keyword{location,}
\keyword{merge,}
\keyword{recordings}
