% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal.R
\name{remove.files}
\alias{remove.files}
\title{Remove Recording Files}
\usage{
remove.files(starting.data, files.remove)
}
\arguments{
\item{starting.data}{A starting data.frame for which you want to remove recordings from. Must have a file.path column.}

\item{files.remove}{A data.frame that contains files that you want remove from the starting data.frame. Must have a file.path column.}
}
\description{
Removes matching files from one data.frame based on input dataframe. Useful to remove already processed recordings or recordings selected already during random file selection.
}
\examples{
remove.files(starting.data, files.remove)
}
\keyword{remove}
