% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal.R
\name{remove.selection}
\alias{remove.selection}
\title{Remove Selected Categories Per Location}
\usage{
remove.selection(starting.data, selection.remove)
}
\arguments{
\item{starting.data}{A starting data.frame for which you want to remove recordings from. Must have a location and categorty column.}

\item{selection.remove}{A data.frame that contains selections you want to remove from the starting data.frame. Must have a location and categorty column.}
}
\description{
This function removes all files that correspond to already selected files so that you can continue with file selction.
}
\examples{
remove.selection(starting.data, selection.remove)
}
\keyword{remove}
