% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveselection.file.R
\name{saveselection.file}
\alias{saveselection.file}
\title{Save recording selection data file to csv}
\usage{
saveselection.file(data, out.path = file.path(getwd(), "SelectionData"))
}
\arguments{
\item{data}{Recording selection data.frame}

\item{out.path}{String path to where to save your file. Defaults to a folder "SelectionData".}
}
\value{
A .csv
}
\description{
Save recording selection data file to csv
}
\examples{
saveselection.file(roundOne,out.path=file.path(getwd(),"SelectionData")
}
\keyword{csv,}
\keyword{recordings}
