% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveselectionfile.R
\name{saveselection}
\alias{saveselection}
\title{Save recording selection data file to csv}
\usage{
saveselection(data, out.path = file.path(getwd(), "SelectionData"))
}
\arguments{
\item{data}{Recording selection data.frame}

\item{out.path}{String path to where to save your file. Defaults to a folder "SelectionData".}
}
\value{
A .csv
}
\description{
Save recording selection data file to csv
}
\keyword{csv,}
\keyword{recordings}
