% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setsoxexe.R
\name{setsox.exe}
\alias{setsox.exe}
\title{Set location of sox.exe}
\usage{
setsox.exe(file.path = NULL)
}
\arguments{
\item{file.path}{A string of the file path to sox.exe. If left blank a file dialog will appear to select the location.}
}
\description{
This function sets and saves the file path to the sox.exe which is needed for many TrillR functions.
}
\examples{
setsox.exe()
}
\keyword{sox.exe}
