% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soxclip.R
\name{sox.clip}
\alias{sox.clip}
\title{Clip wav to desired length}
\usage{
sox.clip(
  file.path,
  out.path = file.path(getwd(), "Clipped_Selection"),
  duration = list(start = 0, end = 180)
)
}
\arguments{
\item{file.path}{String path to .wav file.}

\item{out.path}{String path to where to save your file. Defaults to a folder "Clipped_Selection".}

\item{duration}{A list for the start and end time of the clip in seconds. Default is start = 0 and end = 180 for a 3 min clip.}
}
\value{
A clipped .wav file.
}
\description{
This function uses SoX to clip a .wav file to a desired length.
}
\examples{
sox.clip("C:/LV-01-01-01/LV-01-01-01_20170609_033500.wav",out.path = file.path(getwd(),"Clipped_Selection"), duration = list(start = 0, end = 180))
}
\keyword{categorize,}
\keyword{get,}
\keyword{recordings}
\keyword{wav,}
