% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soxclip.R
\name{sox.clips}
\alias{sox.clips}
\title{Clip muliple wavs to desired length}
\usage{
sox.clips(
  data,
  out.path = file.path(getwd(), "Clipped_Selection"),
  duration = list(start = 0, end = 180)
)
}
\arguments{
\item{data}{String path to .wav file.}

\item{out.path}{String path to where to save your file. Defaults to a folder "Clipped_Selection".}

\item{duration}{A list for the start and end time of the clip in seconds. Default is start = 0 and end = 180 for a 3 min clip.}
}
\value{
Clipped .wav files.
}
\description{
This function uses SoX to clip multiple wav files to a desired length.
}
\examples{
sox.clips(data,out.path = file.path(getwd(),"Spectrograms"), duration = list(start = 0, end = 180))
}
\keyword{clip,}
\keyword{recordings}
\keyword{wav,}
